Reuse `splitFilePath` function
