+++ app/src/lib/utils/filePath.ts
@@ -1,0 +1,9 @@
+export function splitFilePath(filepath: string): { filename: string; path: string } {
+	const parts = filepath.split('/');
+	if (parts.length === 0) return { filename: '', path: '' };
+
+	const filename = parts[parts.length - 1];
+	const path = parts.slice(0, -1).join('/');
+
+	return { filename, path };
+}
